+++ tests/Jroc.Tests/Node/Process/JavaScript/Process_Stdout_Stderr_Write.js
@@ -1,0 +1,6 @@
+"use strict";
+
+process.stdout.write("stdout");
+process.stdout.write(" ok\n");
+process.stderr.write("stderr");
+process.stderr.write(" ok\n");